requirements.txt
appdirs==1.4.4
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-anymail==8.2
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-disqus==0.5
django-environ==0.4.5
django-heroku==0.3.1
django-imagefield==0.13.0
django-js-asset==1.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mock==4.0.3
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
python-decouple==3.4
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.4
whitenoise==5.2.0
